フォントがインストール済であることを検知する

+++ run_once_03_misc.sh
@@ -5,11 +5,16 @@ LANG=C xdg-user-dirs-update --force
 rm -rf デスクトップ ダウンロード テンプレート 公開 ドキュメント ミュージック ピクチャ ビデオ
 
 # HackGen NF 導入
-curl -sLJO https://github.com/yuru7/HackGen/releases/download/v2.10.0/HackGen_NF_v2.10.0.zip
-unzip HackGen_NF_v2.10.0.zip
-mkdir $HOME/.local/share/fonts
-mv HackGen_NF_v2.10.0 $HOME/.local/share/fonts
-rm HackGen_NF_v2.10.0.zip
+if ls $HOME/.local/share/fonts/HackGen_NF_* > /dev/null 2>&1
+then
+  echo "HackGen_NF はインストール済です"
+else
+  curl -sLJO https://github.com/yuru7/HackGen/releases/download/v2.10.0/HackGen_NF_v2.10.0.zip
+  unzip HackGen_NF_v2.10.0.zip
+  mkdir $HOME/.local/share/fonts
+  mv HackGen_NF_v2.10.0 $HOME/.local/share/fonts
+  rm HackGen_NF_v2.10.0.zip
+fi
 
 # 外付けドライブマウントポイント作成
 sudo mkdir /mnt/WDS100T2B0C-EC
